Cleanup of ListClipper code - moved ListClipper C++ code to dedicated files - separated Begin() and BeginV() functions - added & refined documentation
