+++ imguiWrapper.cpp
@@ -194,33 +194,6 @@ void iggCalcTextSize(const char* text, int length, IggBool hide_text_after_doubl
     exportValue(*value, ImGui::CalcTextSize(text, text + length, hide_text_after_double_hash, wrap_width));
 }
 
-IggBool iggListClipperStep(IggListClipper *clipper)
-{
-    ImGuiListClipper imguiClipper;
-    importValue(imguiClipper, *clipper);
-    IggBool returnValue = imguiClipper.Step() ? 1 : 0;
-    exportValue(*clipper, imguiClipper);
-    // needs to be done to prevent assert fail, we don't call end because the cursor will move.
-    imguiClipper.ItemsCount = -1;
-    return returnValue;
-}
-
-void iggListClipperBegin(IggListClipper *clipper, int items_count, float items_height)
-{
-    ImGuiListClipper imguiClipper(items_count, items_height);
-    exportValue(*clipper, imguiClipper);
-    // needs to be done to prevent assert fail, we don't call end because the cursor will move.
-    imguiClipper.ItemsCount = -1;
-}
-
-void iggListClipperEnd(IggListClipper *clipper)
-{
-    ImGuiListClipper imguiClipper;
-    importValue(imguiClipper, *clipper);
-    imguiClipper.End();
-    exportValue(*clipper, imguiClipper);
-}
-
 void iggPushItemWidth(float width)
 {
    ImGui::PushItemWidth(width);
